+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,5 +1,6 @@
 ﻿### Version 3.3.39
 
+- Added: Added support for XMLDocComments on `Method` and `Constructor` `Parameter`s.
 - Added: `ProtectedSetter` added to `CSharpProperty`.
 - Added: `IsExplicitImplementationFor(string @interface)` added to `CSharpMethod` for being able to make a method explicitly implement an interface's method.
 - Added: It is now possible to specify that `CSharpStatementBlock`s and derived types are *not* `SeparatedFromPrevious` by using the the `.SeparatedFromPrevious(false)` method.